+++ precompiles/gov/query_test.go
@@ -34,36 +34,64 @@ var (
 func (s *PrecompileTestSuite) TestGetVotes() {
 	var ctx sdk.Context
 	method := s.precompile.Methods[gov.GetVotesMethod]
+	gas := uint64(200_000)
 	testCases := []struct {
-		name     string
-		malleate func() []gov.WeightedVote
-		args     []interface{}
-		expPass  bool
-		expTotal uint64
-		gas      uint64
+		name        string
+		malleate    func() []gov.WeightedVote
+		args        []interface{}
+		expPass     bool
+		errContains string
+		expTotal    uint64
 	}{
 		{
 			name: "valid query",
 			malleate: func() []gov.WeightedVote {
-				err := s.network.App.GovKeeper.AddVote(s.network.GetContext(), 1, s.keyring.GetAccAddr(0), []*govv1.WeightedVoteOption{{Option: govv1.OptionYes, Weight: "1.0"}}, "")
-				s.Require().NoError(err)
-				return []gov.WeightedVote{
-					{ProposalId: 1, Voter: s.keyring.GetAddr(0), Options: []gov.WeightedVoteOption{{Option: uint8(govv1.OptionYes), Weight: "1.0"}}},
+				proposalID := uint64(1)
+				voter := s.keyring.GetAccAddr(0)
+				voteOption := &govv1.WeightedVoteOption{
+					Option: govv1.OptionYes,
+					Weight: "1.0",
 				}
+
+				err := s.network.App.GovKeeper.AddVote(
+					s.network.GetContext(),
+					proposalID,
+					voter,
+					[]*govv1.WeightedVoteOption{voteOption},
+					"",
+				)
+				s.Require().NoError(err)
+
+				return []gov.WeightedVote{{
+					ProposalId: proposalID,
+					Voter:      s.keyring.GetAddr(0),
+					Options: []gov.WeightedVoteOption{
+						{
+							Option: uint8(voteOption.Option), //nolint:gosec // G115 -- integer overflow is not happening here
+							Weight: voteOption.Weight,
+						},
+					},
+				}}
 			},
 			args:     []interface{}{uint64(1), query.PageRequest{Limit: 10, CountTotal: true}},
 			expPass:  true,
 			expTotal: 1,
-			gas:      200_000,
 		},
 		{
-			name:    "invalid proposal ID",
-			args:    []interface{}{uint64(0), query.PageRequest{Limit: 10, CountTotal: true}},
-			expPass: false,
-			gas:     200_000,
-			malleate: func() []gov.WeightedVote {
-				return []gov.WeightedVote{}
-			},
+			name:        "invalid proposal ID",
+			args:        []interface{}{uint64(0), query.PageRequest{Limit: 10, CountTotal: true}},
+			expPass:     false,
+			errContains: "proposal id can not be 0",
+		},
+		{
+			name:        "fail - invalid number of args",
+			args:        []interface{}{},
+			errContains: fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 2, 0),
+		},
+		{
+			name:        "fail - invalid arg types",
+			args:        []interface{}{"string argument 1", 2},
+			errContains: "error while unpacking args to VotesInput",
 		},
 	}
 
@@ -72,9 +100,13 @@ func (s *PrecompileTestSuite) TestGetVotes() {
 			s.SetupTest()
 			ctx = s.network.GetContext()
 
-			votes := tc.malleate()
+			var votes []gov.WeightedVote
+			if tc.malleate != nil {
+				votes = tc.malleate()
+			}
+
 			var contract *vm.Contract
-			contract, ctx = testutil.NewPrecompileContract(s.T(), ctx, s.keyring.GetAddr(0), s.precompile, tc.gas)
+			contract, ctx = testutil.NewPrecompileContract(s.T(), ctx, s.keyring.GetAddr(0), s.precompile, gas)
 
 			bz, err := s.precompile.GetVotes(ctx, &method, contract, tc.args)
 
@@ -86,6 +118,7 @@ func (s *PrecompileTestSuite) TestGetVotes() {
 				s.Require().Equal(tc.expTotal, out.PageResponse.Total)
 			} else {
 				s.Require().Error(err)
+				s.Require().ErrorContains(err, tc.errContains)
 			}
 		})
 	}
@@ -93,71 +126,99 @@ func (s *PrecompileTestSuite) TestGetVotes() {
 
 func (s *PrecompileTestSuite) TestGetVote() {
 	var voter sdk.AccAddress
+	var voterAddr common.Address
+
 	method := s.precompile.Methods[gov.GetVoteMethod]
+
 	testCases := []struct {
 		name          string
-		malleate      func()
-		propNumber    uint64
+		malleate      func() []interface{}
 		expPass       bool
 		expPropNumber uint64
 		expVoter      common.Address
-		gas           uint64
 		errContains   string
 	}{
 		{
 			name: "valid query",
-			malleate: func() {
+			malleate: func() []interface{} {
 				err := s.network.App.GovKeeper.AddVote(s.network.GetContext(), 1, voter, []*govv1.WeightedVoteOption{{Option: govv1.OptionYes, Weight: "1.0"}}, "")
 				s.Require().NoError(err)
+
+				return []interface{}{uint64(1), voterAddr}
 			},
-			propNumber:    uint64(1),
 			expPropNumber: uint64(1),
 			expVoter:      common.BytesToAddress(voter.Bytes()),
 			expPass:       true,
-			gas:           200_000,
 		},
 		{
-			name:       "invalid proposal ID",
-			propNumber: uint64(10),
-			expPass:    false,
-			gas:        200_000,
-			malleate: func() {
+			name:    "invalid proposal ID",
+			expPass: false,
+			malleate: func() []interface{} {
 				err := s.network.App.GovKeeper.AddVote(s.network.GetContext(), 1, voter, []*govv1.WeightedVoteOption{{Option: govv1.OptionYes, Weight: "1.0"}}, "")
 				s.Require().NoError(err)
+
+				return []interface{}{uint64(10), voterAddr}
 			},
 			errContains: "not found for proposal",
 		},
 		{
-			name:        "non-existent vote",
-			propNumber:  uint64(1),
+			name: "non-existent vote",
+			malleate: func() []interface{} {
+				return []interface{}{uint64(1), voterAddr}
+			},
 			expPass:     false,
-			gas:         200_000,
-			malleate:    func() {},
 			errContains: "not found for proposal",
+		},
+		{
+			name: "invalid number of args",
+			malleate: func() []interface{} {
+				return []interface{}{}
+			},
+			errContains: fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 2, 0),
+		},
+		{
+			name: "fail - invalid proposal id",
+			malleate: func() []interface{} {
+				return []interface{}{"string argument 1", 2}
+			},
+			errContains: "invalid proposal id",
+		},
+		{
+			name: "fail - invalid voter address",
+			malleate: func() []interface{} {
+				return []interface{}{uint64(0), 2}
+			},
+			errContains: "invalid voter address",
 		},
 	}
 
 	for _, tc := range testCases {
 		s.Run(tc.name, func() {
 			s.SetupTest()
+
 			voter = s.keyring.GetAccAddr(0)
+			voterAddr = s.keyring.GetAddr(0)
+			gas := uint64(200_000)
 
-			tc.malleate()
+			var args []interface{}
+			if tc.malleate != nil {
+				args = tc.malleate()
+			}
 
-			contract, ctx := testutil.NewPrecompileContract(s.T(), s.network.GetContext(), s.keyring.GetAddr(0), s.precompile, tc.gas)
+			contract, ctx := testutil.NewPrecompileContract(s.T(), s.network.GetContext(), voterAddr, s.precompile, gas)
 
-			args := []interface{}{tc.propNumber, common.BytesToAddress(voter.Bytes())}
 			bz, err := s.precompile.GetVote(ctx, &method, contract, args)
 
 			expVote := gov.WeightedVote{
 				ProposalId: tc.expPropNumber,
-				Voter:      common.BytesToAddress(voter.Bytes()),
+				Voter:      voterAddr,
 				Options:    []gov.WeightedVoteOption{{Option: uint8(govv1.OptionYes), Weight: "1.0"}},
 				Metadata:   "",
 			}
 
 			if tc.expPass {
 				s.Require().NoError(err)
+
 				var out gov.VoteOutput
 				err = s.precompile.UnpackIntoInterface(&out, gov.GetVoteMethod, bz)
 
@@ -168,7 +229,7 @@ func (s *PrecompileTestSuite) TestGetVote() {
 				s.Require().Equal(expVote.Metadata, out.Vote.Metadata)
 			} else {
 				s.Require().Error(err)
-				s.Require().Contains(err.Error(), tc.errContains)
+				s.Require().ErrorContains(err, tc.errContains)
 			}
 		})
 	}
@@ -580,6 +641,48 @@ func (s *PrecompileTestSuite) TestGetProposals() {
 				err = s.precompile.UnpackIntoInterface(&out, gov.GetProposalsMethod, bz)
 				s.Require().NoError(err)
 				tc.postCheck(out.Proposals, &out.PageResponse)
+			}
+		})
+	}
+}
+
+func (s *PrecompileTestSuite) TestGetParams() {
+	testCases := []struct {
+		name        string
+		malleate    func() []interface{}
+		expPass     bool
+		errContains string
+	}{
+		{
+			"fail - not empty input args",
+			func() []interface{} {
+				return []interface{}{""}
+			},
+			false,
+			fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 0, 1),
+		},
+		{
+			"success - get all params",
+			func() []interface{} {
+				return []interface{}{}
+			},
+			true,
+			"",
+		},
+	}
+
+	for _, tc := range testCases {
+		s.Run(tc.name, func() {
+			s.SetupTest()
+
+			method := s.precompile.Methods[gov.GetParamsMethod]
+			_, err := s.precompile.GetParams(s.network.GetContext(), &method, nil, tc.malleate())
+
+			if tc.expPass {
+				s.Require().NoError(err)
+			} else {
+				s.Require().Error(err)
+				s.Require().Contains(err.Error(), tc.errContains)
 			}
 		})
 	}